pyproject.toml
[project]
name = "gemma3npc"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "google-genai",
    "python-dotenv",
]
